Move middleware log funcs to utils
